rename nycmesh -> meshapi package
